# api/openapi.yaml
openapi: 3.1.0
info:
  title: scrim hub machine API
  version: 0.6.0
  summary: HTTP control surface a remote MCP client (scrim mcp --hub) uses to author canvas content on a scrim hub.
  description: |-
    The machine API exposed by `scrim hub` (see the repo `CLAUDE.md` "Hub" section). It is the
    same serving engine as `scrim serve`, plus a bearer-gated write/read surface for a remote
    client with no shared disk: per-file read/write/edit, whole-canvas push and copy, snapshot
    control, and canvas CRUD.

    **Mostly hub-only.** Every `push`/`files`/`copy`/`snapshots`/`grants`/`tokens`/`claim`/
    `principals` path here, plus `GET /api/openapi.yaml` and `/healthz`, is registered ONLY when the
    daemon is started as a hub (`NewHub`). Five are NOT: `GET /api/status`, `GET`/`POST
    /api/canvases`, `DELETE /api/canvases/{id}`, and `POST /api/stop` are registered
    unconditionally, so the default local daemon (`scrim serve`) exposes them too — gated there by a
    query-string capability token instead of a bearer. This document describes the hub.

    **Authorization model** (`hubgate.go`). A machine bearer presented as `Authorization: Bearer
    <token>` authorizes the machine API, reads included. Two credentials qualify: the ADMIN PUSH
    TOKEN, which authorizes ANY method and is the sole credential `POST /api/stop` accepts; and a
    USER TOKEN, which authorizes the same API bounded to canvases its owner may write (`403`
...
    browser gate.

...

...

...

  license:
...
servers:
...
security:
  - pushToken: []
paths:
...
  /api/canvases:
    get:
      tags: [canvases]
      operationId: listCanvases
      summary: List canvases visible to the caller.
      description: >-
        Lists canvases. On an OIDC hub the list is private by default — only canvases the caller
        owns or has been granted (or, with a `?k=` share-link secret, that link's canvas) are
        returned; the admin push token sees all. On a non-OIDC hub / the local daemon every canvas
        is returned.
      parameters:
        - $ref: '#/components/parameters/LinkSecret'
      responses:
        '200':
          description: Every canvas, each as a Canvas object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Canvas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [canvases]
      operationId: createCanvas
      summary: Create an empty canvas.
      description: >-
        Registers a new, empty canvas directory. The client typically calls this before writing
        files into it (PUT/PATCH return 404 for a canvas that does not exist).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCanvasRequest'
      responses:
        '201':
          description: Canvas created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canvas'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/stop:
    post:
      tags: [status]
      operationId: stopDaemon
      summary: Stop the daemon.
      description: >-
        Acknowledges the request, then triggers a graceful shutdown. The one ADMIN-ONLY operation:
        only the hub's admin push token is accepted. A valid USER token is always `403` — never
        `401` — regardless of what it owns; a session or an anonymous caller is `401`.
      responses:
        '200':
          description: Shutdown acknowledged; the daemon then exits.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                required: [ok]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >-
            A valid user token (or forwarded actor) presented on an admin-only operation. Always
            `403` for a user token, whatever it owns. Plain-text body (gate-level denial).
          content:
            text/plain:
              schema:
                type: string
...
  /api/canvases/{id}/grants:
    parameters:
      - $ref: '#/components/parameters/CanvasId'
    get:
      tags: [grants]
      operationId: listGrants
      summary: List a canvas's owner and sharing grants.
      description: >-
        Returns the canvas's owner and its current view-only grants, secret-free — a share-link
        grant appears only by its public `link_id`, never its secret or hash. The ACL is more
        privileged than the canvas view: only the owner, the admin push token, a verified
        gateway-forwarded actor that owns the canvas, or a principal that is itself an explicit
        `user`/`group` grantee may read it. A caller who can see the canvas only via a share link or
        an `everyone` grant is refused (`403`), so a link holder cannot enumerate the owner's email
        or the other grantees.
      parameters:
        - $ref: '#/components/parameters/LinkSecret'
      responses:
        '200':
          description: The canvas's owner and grants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [grants]
      operationId: createGrant
      summary: Add a sharing grant to a canvas.
      description: >-
        Adds a view-only grant. Only the canvas owner, the admin push token, or a verified
        gateway-forwarded actor that owns the canvas may add a grant; a user token is additionally bounded
        by its `allowed_grant_targets` allowance (`403` when a target is not permitted). For a `link`
        grant the hub mints a fresh link id + secret and returns the raw secret ONCE — append it to
        the canvas URL as `?k=<secret>`. Adding an already-present user/group/everyone grant is a
        no-op (`200`). A browser session that OWNS the canvas may call this directly (the share
        dialog does), unrestricted over its own canvas; `SameSite=Lax` is what makes that CSRF-safe.
      security:
        - pushToken: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGrantRequest'
      responses:
        '200':
          description: The grant already existed (idempotent add).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGrantResponse'
        '201':
          description: Grant added; a link grant's secret is shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGrantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/canvases/{id}/grants/{grantRef}:
    delete:
      tags: [grants]
      operationId: deleteGrant
      summary: Remove a sharing grant.
      description: >-
        Removes matching grants. `grantRef` addresses a link grant by its public `link_id`, a
        user/group grant by its `target`, and the everyone grant by the literal `everyone`. Owner or
        admin only. No matching grant is `404`. A browser session that OWNS the canvas may call this
        directly (the share dialog does); `SameSite=Lax` is what makes that CSRF-safe.
      security:
        - pushToken: []
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CanvasId'
        - name: grantRef
          in: path
          required: true
          schema:
            type: string
          description: The grant reference (link id, target, or "everyone").
      responses:
        '204':
          description: Grant(s) removed. No body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/tokens:
    get:
      tags: [tokens]
      operationId: listTokens
      summary: List the caller's user tokens.
      description: >-
        Lists the calling principal's own user tokens (scoped by principal email), never their
        hashes or raw secrets. On an OIDC hub the caller is a browser session or a user-token
        principal; the admin push token has no email of its own and lists nothing here.
      security:
        - pushToken: []
        - sessionCookie: []
      responses:
        '200':
          description: The caller's tokens (secret-free).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TokenMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [tokens]
      operationId: createToken
      summary: Mint a user token.
      description: >-
        Mints a named user bearer token that acts AS its owner on the machine plane. Authorized for
        a logged-in browser session (mints for itself) or the admin push token (mints for a named
        `owner_email`); a user-token principal is rejected (`403`, no privilege escalation). The raw
        secret is returned ONCE in the response and never again — only its SHA-256 hash is stored.
        `auto_share` grants are applied to canvases the token creates; `allowed_grant_targets`
        bounds what it may later share (enforced by the grant endpoints).
      security:
        - pushToken: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        '201':
          description: Token minted; the raw secret is shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MintTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    pushToken:
      type: http
      scheme: bearer
      description: >-
        A machine bearer presented as `Authorization: Bearer <token>`. Two credentials share this
        scheme. The hub's ADMIN PUSH TOKEN (compared in constant time) authorizes ANY method —
        writes and reads alike, and it is the only credential `POST /api/stop` accepts. A USER
        TOKEN, minted by a logged-in principal, authorizes the same machine API bounded to canvases
        its owner may write (`403` otherwise) and, under OIDC, may read only what its owner may
        view. Absent any valid bearer, a machine-API write is `401`; a read falls to the browser
        gate (OIDC session, or CIDR allowlist + optional read token) and may instead be `403`
        (outside the allowlist) or `401` (OIDC configured, no session).
    sessionCookie:
      type: apiKey
      in: cookie
      name: scrim_session
      description: >-
        The browser session cookie minted by the OIDC login flow (`HttpOnly`, `SameSite=Lax`). The
        session plane is read-only for the machine API, with three deliberate exceptions that accept
        it as a WRITE credential: canvas claim, `/api/tokens*`, and mutating a canvas's own grants.
        Those are owner actions the browser must perform natively; `SameSite=Lax` is what makes them
        CSRF-safe. `POST /api/tokens` accepts a session or the admin push token but rejects a USER
        token with `403` — minting from a token would be privilege escalation.
...
  responses:
...
    Unauthorized:
      description: >-
        Missing or invalid credential on a write, or, for a read, an OIDC-configured hub with no
        session. ALWAYS produced by the gate, so the body is `text/plain` — never the JSON `Error`
        shape.
      content:
        text/plain:
          schema:
            type: string
...